requirements.txt
flask==3.0.2
pillow==10.2.0
numpy==1.26.4
tensorflow==2.15.0
gunicorn==21.2.0
python-dotenv==1.0.1
werkzeug==3.0.1 